requirements.txt
Django
easy-thumbnails
python-magic
django-environ
django-simple-captcha
django-guardian